revert pathbuilder change due to circular dependency
