+++ flood_adapt/object_model/hazard/event/eventset.py
@@ -1,11 +1,11 @@
 from pathlib import Path
 from typing import Any
 
-from flood_adapt.object_model.interface.database import (
-    ObjectDir,
-)
 from flood_adapt.object_model.interface.events import EventSetModel
 from flood_adapt.object_model.interface.object_model import IObject
+from flood_adapt.object_model.interface.path_builder import (
+    ObjectDir,
+)
 
 
 class EventSet(IObject[EventSetModel]):
