+++ flood_adapt/object_model/interface/benefits.py
@@ -5,10 +5,10 @@ from typing import Optional
 import pandas as pd
 from pydantic import BaseModel
 
-from flood_adapt.object_model.interface.database import (
+from flood_adapt.object_model.interface.object_model import IObject, IObjectModel
+from flood_adapt.object_model.interface.path_builder import (
     ObjectDir,
 )
-from flood_adapt.object_model.interface.object_model import IObject, IObjectModel
 
 
 class CurrentSituationModel(BaseModel):
