+++ flood_adapt/object_model/interface/events.py
@@ -3,10 +3,10 @@ from typing import Optional, TypeVar
 
 from pydantic import BaseModel
 
-from flood_adapt.object_model.interface.database import (
+from flood_adapt.object_model.interface.object_model import IObject, IObjectModel
+from flood_adapt.object_model.interface.path_builder import (
     ObjectDir,
 )
-from flood_adapt.object_model.interface.object_model import IObject, IObjectModel
 from flood_adapt.object_model.io.unitfulvalue import (
     UnitfulDirection,
     UnitfulDischarge,
