+++ flood_adapt/object_model/interface/measures.py
@@ -3,10 +3,10 @@ from typing import Any, Generic, Optional, TypeVar
 
 from pydantic import Field, model_validator, validator
 
-from flood_adapt.object_model.interface.database import (
+from flood_adapt.object_model.interface.object_model import IObject, IObjectModel
+from flood_adapt.object_model.interface.path_builder import (
     ObjectDir,
 )
-from flood_adapt.object_model.interface.object_model import IObject, IObjectModel
 from flood_adapt.object_model.io.unitfulvalue import (
     UnitfulDischarge,
     UnitfulHeight,
