+++ flood_adapt/object_model/interface/object_model.py
@@ -7,7 +7,7 @@ import tomli
 import tomli_w
 from pydantic import BaseModel, Field
 
-from flood_adapt.object_model.interface.database import (
+from flood_adapt.object_model.interface.path_builder import (
     ObjectDir,
 )
