+++ flood_adapt/object_model/interface/projections.py
@@ -2,10 +2,10 @@ from typing import Optional
 
 from pydantic import BaseModel
 
-from flood_adapt.object_model.interface.database import (
+from flood_adapt.object_model.interface.object_model import IObject, IObjectModel
+from flood_adapt.object_model.interface.path_builder import (
     ObjectDir,
 )
-from flood_adapt.object_model.interface.object_model import IObject, IObjectModel
 from flood_adapt.object_model.io.unitfulvalue import (
     UnitfulLength,
     UnitfulLengthRefValue,
