+++ flood_adapt/object_model/interface/scenarios.py
@@ -1,9 +1,9 @@
 from abc import abstractmethod
 
-from flood_adapt.object_model.interface.database import (
+from flood_adapt.object_model.interface.object_model import IObject, IObjectModel
+from flood_adapt.object_model.interface.path_builder import (
     ObjectDir,
 )
-from flood_adapt.object_model.interface.object_model import IObject, IObjectModel
 
 
 class ScenarioModel(IObjectModel):
