+++ flood_adapt/object_model/interface/site.py
@@ -3,10 +3,10 @@ from typing import Any, Optional, Union
 
 from pydantic import BaseModel, model_validator
 
-from flood_adapt.object_model.interface.database import (
+from flood_adapt.object_model.interface.object_model import IObject, IObjectModel
+from flood_adapt.object_model.interface.path_builder import (
     ObjectDir,
 )
-from flood_adapt.object_model.interface.object_model import IObject, IObjectModel
 from flood_adapt.object_model.io.unitfulvalue import (
     UnitfulDischarge,
     UnitfulLength,
